make ItemBlockPad an inner class, since nothing else uses it.
